+++ doc/release_notes.rst
@@ -7,10 +7,10 @@
 Release Notes
 ##########################################
 
-.. Upcoming Release
-.. ================
+Upcoming Release
+================
 
-.. * new features
+* New configuration option ``everywhere_powerplants`` to build conventional powerplants everywhere, irrespective of existing powerplants locations, in the network (https://github.com/PyPSA/pypsa-eur/pull/848). 
 
 PyPSA-Eur 0.9.0 (5th January 2024)
 ==================================